+++ WordPress/src/main/java/org/wordpress/android/ui/publicize/PublicizeListActivity.java
@@ -285,7 +285,7 @@ public class PublicizeListActivity extends AppCompatActivity
         PublicizeAccountChooserDialogFragment dialogFragment = new PublicizeAccountChooserDialogFragment();
         Bundle args = new Bundle();
         args.putString(PublicizeConstants.ARG_CONNECTION_ARRAY_JSON, event.getJSONObject().toString());
-        args.putInt(PublicizeConstants.ARG_SITE_ID, event.getSiteId());
+        args.putLong(PublicizeConstants.ARG_SITE_ID, event.getSiteId());
         args.putString(PublicizeConstants.ARG_SERVICE_ID, event.getServiceId());
         dialogFragment.setArguments(args);
         dialogFragment.show(getSupportFragmentManager(), PublicizeAccountChooserDialogFragment.TAG);
